requirements.txt
uvicorn>=0.24.0
pandas>=2.0.0
pillow>=10.0.0
fastapi>=0.104.0
python-multipart>=0.0.6
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
numpy>=1.21.0
matplotlib>=3.5.0
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84